requirements.txt
numpy==1.24.4
streamlit>=1.22.0
pandas>=1.5.3
requests>=2.28.2
python-dotenv>=1.0.0
boto3>=1.26.0
plotly>=5.14.0
streamlit-ace>=0.1.1
opensearch-py>=2.0.0
elasticsearch>=6.8.0,<7.0.0
botocore>=1.29.0
strands-agents>=1.0.0
mcp>=1.0.0
